Used Layout as wrapper for SplitBills in Credit, updated the file paths
