+++ src/components/Splits/Credit/SplitBills.js
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
-import * as actionCreators from '../../state/actionCreators';
+import * as actionCreators from '../../../state/actionCreators';
 import SplitBillsCard from './SplitBillsCard';
 import styled from 'styled-components';
-import Layout from '../Layout/';
+
 export const SplitBills = props => {
   // console.log(props)
   const { lumpState, getUserDetails, getUsers } = props;
@@ -13,19 +13,17 @@ export const SplitBills = props => {
     getUserDetails();
     getUsers();
   }, []);
-  return (
-    <Layout>
-      <StyledDiv>
-        {lumpState.currentUser.bills.map(bill => (
-          <SplitBillsCard
-            key={bill.id}
-            setOwingUsers={setOwingUsers}
-            feature={bill}
-            owingUsers={owingUsers}
-          />
-        ))}
-      </StyledDiv>
-    </Layout>
+  return (    
+    <StyledDiv>
+      {lumpState.currentUser.bills.map(bill => (
+        <SplitBillsCard
+          key={bill.id}
+          setOwingUsers={setOwingUsers}
+          feature={bill}
+          owingUsers={owingUsers}
+        />
+      ))}
+    </StyledDiv>    
   );
 };
