+++ src/components/Splits/Credit/SplitBillsCard.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import AxiosAuth from '../../axios/AxiosAuth';
-import UserComponent from '../UsersComponent';
+import AxiosAuth from '../../../axios/AxiosAuth';
+import UserComponent from '../../UsersComponent';
 import styled from "styled-components";
 
 import {  Card } from "@material-ui/core";
