+++ src/components/Splits/Credit/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import SplitBills from './SplitBills';
+import Layout from '../../Layout';
+
+export default () => {
+    return (
+        <Layout title='My Credit'>
+            <SplitBills />
+        </Layout>
+    )
+}
